doc_tagger_daemon/requirements.txt
azure-functions==1.20.*
requests>=2.31,<3
httpx>=0.24,<0.28
azure-identity>=1.16,<2
azure-keyvault-secrets>=4.7,<5
azure-storage-blob>=12.18,<13
openai>=1.30,<2
python-docx>=0.8.11,<1
pdfplumber>=0.10,<0.12
itsdangerous>=2.1,<3